unir todas las funciones de jugadores en una clase estatica  llamada PlayerManager y todos las funciones de equipos en una llamada TeamManager y agruparlos en una carpeta llamada 'managers' los archivos database, crear una clase para imprimir diferentes menus y juntarla en una carpeta con la funcion de createId llamada 'utils', limpiar un poco en archivo principal 'main'
